refactor: move endpoint classes to a separate package
* They're associated with web API endpoints, so they don't belong with the pages.
