+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/integration/endpoints/SamlLogoutAuthSourceEndpoint.java
@@ -1,13 +1,11 @@
-package org.cloudfoundry.identity.uaa.integration.pageObjects;
+package org.cloudfoundry.identity.uaa.integration.endpoints;
 
+import org.cloudfoundry.identity.uaa.integration.pageObjects.Page;
+import org.cloudfoundry.identity.uaa.integration.pageObjects.SamlWelcomePage;
 import org.openqa.selenium.WebDriver;
 
-public class SamlLogoutAuthSourceEndpoint extends Page {
+public class SamlLogoutAuthSourceEndpoint {
     static final private String urlPath = "/module.php/core/logout";
-
-    public SamlLogoutAuthSourceEndpoint(WebDriver driver) {
-        super(driver);
-    }
 
     static public SamlWelcomePage logoutAuthSource_goesToSamlWelcomePage(WebDriver driver, String baseUrl, String authSource) {
         driver.get(baseUrl + urlPath + "/" + authSource);
